add lint plugin fast-refresh,
and move non-component imports to own file, to allow HMR
