+++ package.json
@@ -25,6 +25,7 @@
     "devDependencies": {
         "@comet/dev-process-manager": "^2.0.0",
         "dotenv-cli": "^6.0.0",
+        "eslint-plugin-react-refresh": "^0.4.3",
         "husky": "^7.0.1",
         "lint-staged": "^11.1.1",
         "npm-run-all": "^4.1.5",
